# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CompleteRequest:
    properties:
      achievements:
        type: string
      report:
        type: string
    type: object
...
  dto.CreateFormRequest:
    properties:
      description:
        maxLength: 500
        type: string
      fields:
        items:
          $ref: '#/definitions/model.FormField'
        type: array
      name:
        maxLength: 100
        type: string
      status:
        type: integer
    required:
    - name
    type: object
...
  dto.FileUploadResponse:
    properties:
      category:
        type: string
      file_size:
        type: integer
      filename:
        type: string
      id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      original_name:
        type: string
      size:
        type: integer
      thumbnail_url:
        type: string
      uploaded_at:
        type: string
      url:
        type: string
    type: object
  dto.FormDetail:
    properties:
      created_at:
        type: string
      description:
        type: string
      fields:
        items:
          $ref: '#/definitions/model.FormField'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: integer
      submission_count:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.InternshipConfigRequest:
    properties:
      allow_minister_edit:
        type: boolean
      allow_student_edit:
        type: boolean
      ranking_visible:
        type: boolean
    type: object
...
  dto.MentorCommentRequest:
    properties:
      mentor_comment:
        type: string
    required:
    - mentor_comment
    type: object
...
  dto.ReportRequest:
    properties:
      end_time:
        type: string
      format:
        type: string
      module:
        type: string
      start_time:
        type: string
    type: object
...
  dto.SubmitResponse:
    properties:
      submission_id:
        type: string
      submitted_at:
        type: string
    type: object
...
  dto.UpdateFormRequest:
    properties:
      description:
        maxLength: 500
        type: string
      fields:
        items:
          $ref: '#/definitions/model.FormField'
        type: array
      name:
        maxLength: 100
        type: string
      status:
        type: integer
    type: object
...
  model.FieldOption:
    properties:
      children:
        items:
          $ref: '#/definitions/model.FieldOption'
        type: array
      label:
        type: string
      value: {}
    type: object
  model.FieldValidation:
    properties:
      max_length:
        type: integer
      max_value:
        type: number
      message:
        type: string
      min_length:
        type: integer
      min_value:
        type: number
      pattern:
        type: string
    type: object
  model.FormField:
    properties:
      default: {}
      label:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/model.FieldOption'
        type: array
      placeholder:
        type: string
      props:
        additionalProperties: true
        type: object
      required:
        type: boolean
      type:
        type: string
      validation:
        $ref: '#/definitions/model.FieldValidation'
      visible_when:
        $ref: '#/definitions/model.VisibleWhen'
    type: object
  model.VisibleWhen:
    properties:
      field:
        type: string
      operator:
        type: string
      value: {}
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: 高校计算机协会一体化管理平台
  title: StarByte API
  version: "1.0"
paths:
...
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: 单文件上传到对象存储
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 分类
        in: formData
        name: category
        type: string
      - description: 是否公开
        in: formData
        name: is_public
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileUploadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件
  /files/upload-batch:
    post:
      consumes:
      - multipart/form-data
      description: 一次上传多个文件，表单字段名为 files 或 files[]
      parameters:
      - description: 文件列表
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量上传文件
      tags:
      - 文件
  /forms:
    get:
      description: 列出表单定义；无 form:read 时仅返回已发布表单
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 状态 0草稿 1已发布 2已停用
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 表单列表
      tags:
      - 表单
    post:
      consumes:
      - application/json
      description: 创建动态表单定义，需要 form:write
      parameters:
      - description: 表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FormDetail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建表单
      tags:
      - 表单
  /forms/{id}:
    get:
      description: 按 ID 获取表单字段定义
      parameters:
      - description: 表单 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FormDetail'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 表单定义
      tags:
      - 表单
    put:
      consumes:
      - application/json
      description: 更新表单定义或状态，需要 form:write
      parameters:
      - description: 表单 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FormDetail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新表单
      tags:
      - 表单
  /forms/{id}/submissions:
    get:
      description: 列出某表单的提交记录，需要 form:read
      parameters:
      - description: 表单 ID
        in: path
        name: id
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 提交记录
      tags:
      - 表单
  /forms/{id}/submit:
    post:
      consumes:
      - application/json
      description: 提交已发布表单的填写数据，需要 form:submit
      parameters:
      - description: 表单 ID
        in: path
        name: id
        required: true
        type: string
      - description: 字段值
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SubmitResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 提交表单
      tags:
      - 表单
...
  /internships/{id}/complete:
    post:
      consumes:
      - application/json
      description: 完成实习
      parameters:
      - description: 实习ID
        in: path
        name: id
        required: true
        type: string
      - description: 报告与成果
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 完成实习
      tags:
      - 实习
  /internships/{id}/mentor-comment:
    post:
      consumes:
      - application/json
      description: 指导老师评价
      parameters:
      - description: 实习ID
        in: path
        name: id
        required: true
        type: string
      - description: 评价
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MentorCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 指导老师评价
      tags:
      - 实习
  /internships/{id}/report:
    post:
      consumes:
      - application/json
      description: 提交实习报告
      parameters:
      - description: 实习ID
        in: path
        name: id
        required: true
        type: string
      - description: 报告
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 提交实习报告
      tags:
      - 实习
...
  /system/internship-config:
    get:
      description: 获取实习配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取实习配置
      tags:
      - 实习
    put:
      consumes:
      - application/json
      description: 更新实习配置
      parameters:
      - description: 配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.InternshipConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新实习配置
      tags:
      - 实习
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"